+++ analytics/emr-eks-yunikorn/examples/emr-yunikorn-gang-scheduling/driver-pod-template.yaml
@@ -1,0 +1,62 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: ny-taxi-driver
+  namespace: emr-data-team-a
+  labels:
+    applicationId: ny-taxi-yunikorn
+    queue: root.test
+  annotations:
+    yunikorn.apache.org/schedulingPolicyParameters: "placeholderTimeoutSeconds=30 gangSchedulingStyle=Hard"
+    yunikorn.apache.org/task-group-name: "spark-driver"
+    # YuniKorn Gang Scheduling Configuration
+    # minMember should match with driver and executor instances
+    # minResource cpu and memory should match with driver and executor cpu and memory. This includes sidecar container resources.
+    # Resource below should never be less than the actual resources defined for Driver and Executor with sidecar containers
+    yunikorn.apache.org/task-groups: |-
+      [{
+          "name": "spark-driver",
+          "minMember": 1,
+          "minResource": {
+            "cpu": "1200m",
+            "memory": "15Gi"
+          },
+          "nodeSelector": {
+            "NodeGroupType": "SPARK_DRIVER",
+            "eks.amazonaws.com/capacityType": "ON_DEMAND"
+          }
+        },
+        {
+          "name": "spark-executor",
+          "minMember": 6,
+          "minResource": {
+            "cpu": "1200m",
+            "memory": "15Gi"
+          },
+          "nodeSelector": {
+            "NodeGroupType": "SPARK_EXEC_SPOT",
+            "eks.amazonaws.com/capacityType": "SPOT"
+          }
+      }]
+spec:
+  volumes:
+    - name: spark-local-dir-1
+      hostPath:
+        path: /local1  # Formatted and mounted by terraform managed node groups; hadoop permissions applied using bootstrap data
+        type: Directory
+  nodeSelector:
+    NodeGroupType: SPARK_DRIVER
+
+  initContainers:
+    - name: volume-permissions
+      image: public.ecr.aws/y4g4v0z7/busybox
+      command: [ 'sh', '-c', 'chown 999 /local1' ]
+      volumeMounts:
+        - mountPath: "/local1"
+          name: "spark-local-dir-1"
+  containers:
+    - name: spark-kubernetes-driver # Don't change this name. EMR on EKS looking for this name
+      volumeMounts:
+        - name: spark-local-dir-1
+          mountPath: /emrdata
+          readOnly: false
