--- analytics/emr-eks-yunikorn/examples/spark/spark-scripts/pod-templates/spark-executor-team-a.yaml
@@ -1,33 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: ny-taxi-exec
-  namespace: emr-data-team-a
-  labels:
-    applicationId: ny-taxi-yunikorn
-    queue: root.test
-  annotations:
-    yunikorn.apache.org/task-group-name: "spark-executor"
-
-spec:
-  volumes:
-    - name: spark-local-dir-1
-      hostPath:
-        path: /local1  # Formatted and mounted by terraform managed node groups; hadoop permissions applied using bootstrap data
-        type: Directory
-  nodeSelector:
-    NodeGroupType: SPARK_EXEC_SPOT
-
-  initContainers:
-    - name: volume-permissions
-      image: public.ecr.aws/y4g4v0z7/busybox
-      command: [ 'sh', '-c', 'chown 999 /local1' ]
-      volumeMounts:
-        - mountPath: "/local1"
-          name: "spark-local-dir-1"
-  containers:
-    - name: spark-kubernetes-executor # Don't change this name. EMR on EKS looking for this name
-      volumeMounts:
-        - name: spark-local-dir-1
-          mountPath: /emrdata
-          readOnly: false
