--- analytics/emr-eks-yunikorn/examples/spark/spark-scripts/pod-templates/spark-executor-team-b.yaml
@@ -1,20 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: ny-taxi-exec
-  namespace: emr-data-team-b
-spec:
-  volumes:
-    - name: spark-local-dir-1
-      hostPath:
-        path: /local1  # Formatted and mounted by terraform managed node groups; hadoop permissions applied using bootstrap data
-        type: Directory
-  nodeSelector:
-    NodeGroupType: SPARK_EXEC_SPOT
-
-  containers:
-    - name: spark-kubernetes-executor # Don't change this name. EMR on EKS looking for this name
-      volumeMounts:
-        - name: spark-local-dir-1
-          mountPath: /data1
-          readOnly: false
