+++ analytics/emr-eks-yunikorn/main.tf
@@ -99,7 +99,7 @@ module "eks_blueprints" {
       node_group_name = "spark-driver-grp"
       subnet_ids      = [module.vpc.private_subnets[0]]
       instance_types  = ["r5d.4xlarge"] # YuniKorn Gangschueduling Pause pods are not supporting ARM64 instances
-      ami_type        = "AL2_x86_64" # Graviton AL2_ARM_64
+      ami_type        = "AL2_x86_64"    # Graviton AL2_ARM_64
       capacity_type   = "ON_DEMAND"
       # Enable this option only when you are using NVMe disks
       format_mount_nvme_disk = true # Mounts NVMe disks to /local1, /local2 etc. for multiple NVMe disks
@@ -152,7 +152,7 @@ module "eks_blueprints" {
       node_group_name = "spark-exec-spot"
       subnet_ids      = [module.vpc.private_subnets[0]]
       instance_types  = ["r5d.4xlarge", "r5d.8xlarge", "r5d.12xlarge"] # Graviton ["r6gd.4xlarge", "r6gd.8xlarge", "r6gd.12xlarge"]
-      ami_type        = "AL2_x86_64" # Graviton AL2_ARM_64
+      ami_type        = "AL2_x86_64"                                   # Graviton AL2_ARM_64
       capacity_type   = "SPOT"
 
       # Enable this option only when you are using NVMe disks
