+++ analytics/emr-eks-yunikorn/variables.tf
@@ -7,7 +7,7 @@ variable "name" {
 variable "region" {
   description = "region"
   type        = string
-  default     = "us-west-2"
+  default     = "eu-west-1"
 }
 
 variable "eks_cluster_version" {
