--- analytics/spark-k8s-operator/spark-samples/pyspark-pi-job.yaml
@@ -1,36 +1,0 @@
-# NOTE: This example requires the following prerequisites before executing the jobs
-# 1. Ensure spark-team-a name space exists
-
----
-apiVersion: "sparkoperator.k8s.io/v1beta2"
-kind: SparkApplication
-metadata:
-  name: pyspark-pi
-  namespace: spark-team-a
-spec:
-  type: Python
-  pythonVersion: "3"
-  mode: cluster
-  image: "gcr.io/spark-operator/spark-py:v3.1.1"
-  imagePullPolicy: Always
-  mainApplicationFile: local:///opt/spark/examples/src/main/python/pi.py
-  sparkVersion: "3.1.1"
-  restartPolicy:
-    type: OnFailure
-    onFailureRetries: 3
-    onFailureRetryInterval: 10
-    onSubmissionFailureRetries: 5
-    onSubmissionFailureRetryInterval: 20
-  driver:
-    cores: 1
-    coreLimit: "1200m"
-    memory: "512m"
-    labels:
-      version: 3.1.1
-    serviceAccount: spark-team-a
-  executor:
-    cores: 1
-    instances: 2
-    memory: "512m"
-    labels:
-      version: 3.1.1
