--- analytics/spark-k8s-operator/spark-samples/spark-teams-setup.yaml
@@ -1,52 +1,0 @@
-# YAML Config for setting up a new Data team
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: spark-team-a
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: spark-team-a
-  namespace: spark-team-a
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: spark-role
-rules:
-  - apiGroups: [""]
-    resources: ["namespaces"]
-    verbs: ["get"]
-  - apiGroups: [""]
-    resources: ["serviceaccounts", "services", "configmaps", "events", "pods", "pods/log"]
-    verbs: ["get", "list", "watch", "describe", "create", "edit", "delete", "deletecollection", "annotate", "patch", "label"]
-  - apiGroups: [""]
-    resources: ["secrets"]
-    verbs: ["create", "patch", "delete", "watch"]
-  - apiGroups: ["apps"]
-    resources: ["statefulsets", "deployments"]
-    verbs: ["get", "list", "watch", "describe", "create", "edit", "delete", "annotate", "patch", "label"]
-  - apiGroups: ["batch"]
-    resources: ["jobs"]
-    verbs: ["get", "list", "watch", "describe", "create", "edit", "delete", "annotate", "patch", "label"]
-  - apiGroups: ["extensions"]
-    resources: ["ingresses"]
-    verbs: ["get", "list", "watch", "describe", "create", "edit", "delete", "annotate", "patch", "label"]
-  - apiGroups: ["rbac.authorization.k8s.io"]
-    resources: ["roles", "rolebindings"]
-    verbs: ["get", "list", "watch", "describe", "create", "edit", "delete", "deletecollection", "annotate", "patch", "label"]
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: spark-role-binding
-subjects:
-  - kind: ServiceAccount
-    name: spark-team-a
-    namespace: spark-team-a
-roleRef:
-  kind: ClusterRole
-  name: spark-role
-  apiGroup: rbac.authorization.k8s.io
